% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand.R
\name{blockrand}
\alias{blockrand}
\title{Block randomization}
\usage{
blockrand(n, arms = LETTERS[seq(2)], blocksizes = 1:4, pascal = TRUE, ...)
}
\arguments{
\item{n}{total number of randomizations (within a single stratum)}

\item{arms}{number of arms to randomise}

\item{blocksizes}{numbers of each arm to include in blocks}

\item{pascal}{logical, whether to use pascal's triangle to determine block sizes}

\item{...}{arguments passed on to other methods (currently unused)}
}
\value{
a data frame with columns block, blocksize, seq_in_block, arm
}
\description{
Generate a randomization list for a single stratum with blocks of varying sizes.
}
\details{
By default, frequency of the different block sizes is determined using Pascal's
triangle.
This has the advantage that small and large block sizes are less common than
intermediate sized blocks, which helps with making it more difficult to guess
future allocations, and reduces the risk of finishing in the middle of a large
block.

Unbalanced randomization is possible by specifying the same arm label multiple times.
}
\examples{
set.seed(1)
blockrand(10)

# different arm labels
blockrand(10, arms = c("Arm 1", "Arm 2"))

# block sizes 2, 4, and 6, 2 arms
blockrand(20, blocksizes = 1:3)

# unbalanced randomisation (2:1)
blockrand(12, arms = c("Arm 1", "Arm 1", "Arm 2"))

# fixed block sizes
blockrand(10, blocksizes = 2)

}
