% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pascalprops.R
\name{pascalprops}
\alias{pascalprops}
\title{Accessory function to calculate the probabilities of Pascals triangle}
\usage{
pascalprops(n)
}
\arguments{
\item{n}{number of elements in the row}
}
\description{
Accessory function to calculate the probabilities of Pascals triangle
}
