% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rando_balance}
\alias{rando_balance}
\title{rando_balance demonstration dataset}
\format{
\subsection{\code{rando_balance}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{strat1, strat2}{Binary stratification variables}
\item{rando_res}{Balanced randomisation result}
\item{rando_res2}{Unbalanced randomisation result}
}
}
}
\usage{
rando_balance
}
\description{
A synthetic dataset used in examples. The dataset contains two randomisation
result variables (\code{rando_res} and \code{rando_res2}) and two stratification variables (\code{strat1} and
\code{strat2}).
}
\keyword{datasets}
