% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.randolist}
\alias{summary.randolist}
\title{Summary method fro randolist objects}
\usage{
\method{summary}{randolist}(object, ...)
}
\arguments{
\item{object}{randolist object}

\item{...}{additional arguments (currently unused)}
}
\description{
Create a short summary report of the aspects of the randomisation list, which
could be used for quality control.
}
\examples{
r <- randolist(20)
print(summary(r))

r2 <- randolist(20, strata = list(sex = c("M", "F")))
print(summary(r2))

# NOTE: explicitly printing isn't technically necessary

}
