% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand.R
\name{blockrand}
\alias{blockrand}
\title{Block randomization}
\usage{
blockrand(n, arms = LETTERS[seq(2)], blocksizes = 1:4, pascal = TRUE)
}
\arguments{
\item{n}{total number of randomizations (within a single strata)}

\item{arms}{number of arms to randomise}

\item{blocksizes}{vector of numbers of each arm to include in blocks}

\item{pascal}{logical, whether to use pascal's triangle to determine block sizes}
}
\value{
a data frame with columns block, blocksize, seq_in_block, arm
}
\description{
Block randomization
}
\examples{

blockrand(100, blocksizes = c(1, 2))

}
