% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.randolist}
\alias{summary.randolist}
\alias{summary.randostratalist}
\title{Summary method for randolist objects}
\usage{
\method{summary}{randolist}(object, ...)

\method{summary}{randostratalist}(object, ...)
}
\arguments{
\item{object}{a randolist object}

\item{...}{additional arguments (currently unused)}
}
\description{
Summary method for randolist objects
}
\section{Functions}{
\itemize{
\item \code{summary(randostratalist)}: summary.randolist

}}
