% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.randolist}
\alias{summary.randolist}
\title{Summary method fro randolist objects}
\usage{
\method{summary}{randolist}(object, ...)
}
\arguments{
\item{object}{randolist object}

\item{...}{additional arguments (currently unused)}
}
\description{
Create a short summary report of the aspects of the randomisation list.
}
\examples{
r <- randolist(20)
summary(r)

r2 <- randolist(20, strata = list(sex = c("M", "F")))
summary(r2)

}
